hack: change sha256 to sha1 in template-variables.rc
The provider acts weird when it comes to checking the thumbprint. The provider suggests to use sha256 for the thumbprint, but kubectl explain VCluster says that thumbprint expects a sha1 hash. So it's clear there's some conflicting information in the documentation.

Some experimenting needs to be done in regards to removing the insecure-flag and using sha256 or sha1 for the thumbprint.

The error can be seen in vsphere-cpi pods, the logs complain about "thumbprint does not match".

Right now there is no error because of a combination of insecure-flag: true and using sha1. Not sure which one of these solves the issue, but it's solved for now.