Improves how the k8sd proxy client creation process
Currently, there are a few issues with the way we create the k8sd proxy
client, which can lead to the use of an invalid client, leading to
errors when trying to perform k8sd-related operations.

We now make sure that the picked Pod is Ready, and that the node's k8sd API
endpoint is responsive.