fix: include the k8s version in the CP status
CAPI now expects the control plane status to include the minimum
K8s version of the cluster nodes.

If unset, the CAPI prechecks consider that the cluster is not ready
yet and will prevent workload machines from being deployed.

https://github.com/kubernetes-sigs/cluster-api/commit/c8a5f18bc49b4f8b555129b7f5c0471f15f24856
https://github.com/kubernetes-sigs/cluster-api/blob/3cce0d973682f11ab0f0ba1c2522eba66dac2d91/internal/controllers/machineset/machineset_preflight.go#L75-L84
https://github.com/kubernetes-sigs/cluster-api/blob/3cce0d973682f11ab0f0ba1c2522eba66dac2d91/internal/controllers/machineset/machineset_preflight.go#L75-L84

This change will add the CP Version status field and set it
based on the lowest K8s version identified across the cluster
nodes.