test: Assert the number of cluster members on node scaling test (#120)
In the node scaling test, we are asserting the number of control plane
nodes in the workload cluster after scaling the number. However, we do
not check the k8sd cluster to see its view.

There have been cases in which the number of control plane nodes in the
Kubernetes workload cluster does not match the number of nodes in the
k8sd cluster. This will add some assertions, making sure those numbers
match.