Add RollingUpdate MaxSurge
The CK8sControlPlane will have the extra configuration option to set
the maxSurge spec.strategy.rollingUpdate.maxSurge to 0.

spec:
  strategy:
      rollingUpdate:
        maxSurge: 1 | 0
By default the maxSurge is set to value 1, meaning that when a cluster rollingUpdate is performed, the ControlPlanes are rolled out in the N, N+1, N fashion (where N is the number of total Control Planes). This workflow always required one extra spare Machine to be available during the rolling update.

If the maxSurge is set to value 0, meaning that when a cluster rollingUpdate is performed, the ControlPlanes are rolled out in the N, N-1, N fashion (where N is the number of total Control Planes). This workflow does not require one Machine to be available during the rolling update.