Adds bom.json file to the snap docker images (#144)
When joining a Kubernetes cluster, k8sd tries to determine the
Kubernetes version by checking the bom.json file included in the
snap.

The docker template does not include that file, meaning that joining
control plane nodes will fail to join because of this.

The components are now built in the same stage, as it is required by
generate-bom.py.