Add support for CP rollout strategy configuration (#135)
* test: Adds Upgrade test scenario with MaxSurge=0

Currently, we only test the rolling upgrade with the default MaxSurge=1
(for a 3-node control plane cluster, a 4th node is created and joined
before removing one of the nodes).

This adds a test for the rolling upgrade with MaxSurge=0 scenario (
for a 3-node control plane cluster, it first removes a node, then
creates and joins a new one). The test ensures that the numbers of
Machines present in CAPI is only 3 for this scenario.

* Add RollingUpdate MaxSurge

The CK8sControlPlane will have the extra configuration option to set
the maxSurge spec.strategy.rollingUpdate.maxSurge to 0.

spec:
  strategy:
      rollingUpdate:
        maxSurge: 1 | 0
By default the maxSurge is set to value 1, meaning that when a cluster rollingUpdate is performed, the ControlPlanes are rolled out in the N, N+1, N fashion (where N is the number of total Control Planes). This workflow always required one extra spare Machine to be available during the rolling update.

If the maxSurge is set to value 0, meaning that when a cluster rollingUpdate is performed, the ControlPlanes are rolled out in the N, N-1, N fashion (where N is the number of total Control Planes). This workflow does not require one Machine to be available during the rolling update.

---------

Co-authored-by: Claudiu Belu <claudiu.belu@canonical.com>